DRIVERS: clk: Allow clock adjustment back to requested rate after parent changes
This functionality is required by the RK3399 SoC clock driver.

Given a hirarchy of clk1 -> [div] -> clk2, when the rate of clk1 gets
changed, clk2 changes as well, as the divider stays the same. There may
be cases where a user of clk2 needs it at a specific rate, so clk2
needs to be readjusted for the changed rate of clk1.

This patch adds a flag to allow a clock to indicate its rate must be
re-adjusted to its requested rate if the parent clock rate changes.
The flag is checked when changing a clock rate, to make the necessary
re-adjustment.