DRIVERS: pinctrl: Fix compiler warning
The 'struct dentry' parameter in pinconf_init_device_debugfs
produces a compiler warning, as there is no declaration for it.
This patch fixes the warning by changing the conditional compilation
logic. The function is only referenced when CONFIG_DEBUF_FS is
defined, and if it is defined, the stub function then only wanted
when CONFIG_PINCONF is not defined.