CORE: The result from vmm_devtree_getattr is ambiguous.
vmm_devtree_getattr returns NULL if either the attribute is
not found, or if the attribute is present, but empty. This means
it is not possible to tell whether an empty property is present
in a device node. This in turn results in
of_property_read_u32() returning the wrong result when an
empty property is not present, breaking the operation of Linux
drivers (pinctrl in particular).

This patch fixes this by getting a pointer to the attribute
first and then getting the value - returning the correct
result if either is found to be NULL.